Refactor PostProcessingBlock to StringParserBlock and improve generate method architecture
- Rename PostProcessingBlock to StringParserBlock for clarity
- Move StringParserBlock from utilblocks.py to dedicated llm_utils.py module
- Refactor generate method to use single-sample _generate method for better modularity
- Update imports and __all__ exports to reflect new structure
- Update test file location and imports to match new module structure
